BE: update user domain permission; FE: add unauthorization in admin site

+++ src/frontend/src/features/admin/admin-roles/services/role.service.ts
@@ -12,7 +12,7 @@ import {
 
 export const roleService = {
   getRoles: async (params?: RoleGetAllRequest): Promise<RoleGetAllResponse> => {
-    const response = await apiClient.GET("/shop/v1/roles", {
+    const response = await apiClient.GET("/shop/v1/admin/roles", {
       params: {
         query: params,
       },
@@ -24,7 +24,7 @@ export const roleService = {
   },
 
   getRoleById: async (id: string): Promise<RoleGetByIdResponse> => {
-    const response = await apiClient.GET("/shop/v1/roles/{id}", {
+    const response = await apiClient.GET("/shop/v1/admin/roles/{id}", {
       params: {
         path: { id },
       },
@@ -36,7 +36,7 @@ export const roleService = {
   },
 
   createRole: async (body: RoleCreateRequest): Promise<RoleCreateResponse> => {
-    const response = await apiClient.POST("/shop/v1/roles/staff", {
+    const response = await apiClient.POST("/shop/v1/admin/roles/staff", {
       body,
     });
     if (response.error) {
@@ -46,7 +46,7 @@ export const roleService = {
   },
 
   updateRole: async (id: string, body: RoleUpdateRequest): Promise<RoleUpdateResponse> => {
-    const response = await apiClient.PATCH("/shop/v1/roles/{id}", {
+    const response = await apiClient.PATCH("/shop/v1/admin/roles/{id}", {
       params: {
         path: { id },
       },
@@ -59,7 +59,7 @@ export const roleService = {
   },
 
   deleteRole: async (id: string): Promise<RoleDeleteResponse> => {
-    const response = await apiClient.DELETE("/shop/v1/roles/{id}", {
+    const response = await apiClient.DELETE("/shop/v1/admin/roles/{id}", {
       params: {
         path: { id },
       },
